Fix comments from reviews
