+++ dmake/commands/__init__.py
@@ -1,0 +1,3 @@
+from . import stop
+from . import deploy
+from . import release
