+++ dmake/commands/deploy.py
@@ -1,0 +1,6 @@
+from dmake.core import make
+
+
+def entry_point(options):
+    options.dependencies = True
+    make(options)
