+++ dmake/core.py
@@ -1,12 +1,10 @@
-import os, sys
-import re
+import os
+import sys
 import uuid
-import inquirer
 
 import dmake.common as common
-from   dmake.common import DMakeException, SharedVolumeNotFoundException
-from   dmake.deepobuild import DMakeFile, append_command
-import dmake.github as github
+from dmake.common import DMakeException, SharedVolumeNotFoundException
+from dmake.deepobuild import DMakeFile, append_command
 
 tag_push_error_msg = "Unauthorized to push the current state of deployment to git server. If the repository belongs to you, please check that the credentials declared in the DMAKE_JENKINS_SSH_AGENT_CREDENTIALS and DMAKE_JENKINS_HTTP_CREDENTIALS allow you to write to the repository."
 
@@ -595,128 +593,6 @@ def generate_command(file_name, cmds):
 
 def get_tag_name():
     return 'deployed_version_%s' % common.branch
-
-###############################################################################
-
-def stop(options):
-    common.run_shell_command("docker rm -f `docker ps -q -f name=%s.%s.%s`" % (common.repo, common.branch, common.build_id))
-
-def deploy(options):
-    options.dependencies = True
-    make(options)
-
-def release(options):
-    github.check_token()
-
-    app = getattr(options, 'app')
-    branch = options.branch
-    owner = os.getenv('DMAKE_GITHUB_OWNER', None)
-    if owner is None:
-        raise DMakeException("Your need to define your Github account/organization name by setting the DMAKE_GITHUB_OWNER environment variable")
-
-    release_re = re.compile(r'v(\d+)\.(\d+)(?:\.(\d+))?(?:-rc\.(\d+))?')
-
-    def tag_to_key(tag):
-        match = release_re.match(tag)
-        if match is None:
-            return None
-        major = int(match.group(1))
-        minor = int(match.group(2))
-        patch = int(match.group(3)) if match.group(3) is not None else None
-        rc    = int(match.group(4)) if match.group(4) is not None else None
-        return (major, minor, patch, rc)
-
-    def key_to_tag(key):
-        tag = "v{}.{}".format(key[0], key[1])
-        if key[2] is not None:
-            tag += '.{}'.format(key[2])
-        if key[3] is not None:
-            tag += '-rc.{}'.format(key[3])
-        return tag
-
-    release_list = {}
-    for release in github.releases(owner, app):
-        key = tag_to_key(release['tag_name'])
-        if key is not None:
-            release_list[key] = {
-                'tag': release['tag_name'],
-                'name': release['name']
-            }
-    sorted_release_keys = sorted(release_list.keys(), reverse=True)
-    latest_per_major_minor = {}
-    for key in sorted_release_keys:
-        major, minor, patch, rc = key
-        if (major, minor) not in latest_per_major_minor:
-            latest_per_major_minor[(major, minor)] = key
-
-    # Ask for previous version
-    questions = [
-        inquirer.List(
-            'prev',
-            message="Which was the previously released version?",
-            choices=[release_list[key]['tag'] for key in sorted_release_keys if key[0:2] in latest_per_major_minor and latest_per_major_minor[key[0:2]] == key] + ['Other'],
-        ),
-    ]
-    answers = inquirer.prompt(questions)
-    if answers['prev'] == 'Other':
-        questions = [
-            inquirer.List(
-                'prev',
-                message="Which was the previously released version?",
-                choices=[release_list[key]['tag'] for key in sorted_release_keys],
-                carousel=True
-            ),
-        ]
-        answers = inquirer.prompt(questions)
-    prev_version = answers['prev']
-
-    # Ask for next version
-    key = tag_to_key(prev_version)
-    next_tag = [
-        (key[0], key[1] + 1, 0 if key[2] is not None else None, None),
-        (key[0] + 1, 0, 0 if key[2] is not None else None, None)
-    ]
-    if key[2] is not None:
-        next_tag.insert(0, (key[0], key[1], key[2] + 1, None))
-    first = next_tag[0]
-    next_tag.insert(len(next_tag), (first[0], first[1], first[2], first[3] + 1 if first[3] is not None else 1))
-
-    questions = [
-        inquirer.List(
-            'next',
-            message="Which will be the next version?",
-            choices=[key_to_tag(tag) for tag in next_tag],
-        ),
-    ]
-    next_version = inquirer.prompt(questions)['next']
-    prerelease = tag_to_key(next_version)[3] is not None
-
-    # Compute change log
-    common.run_shell_command("git fetch --tags", ignore_error=True)  # git fetch outputs to stderr ?
-    change_log_cmd = "git log {prev}...deployed_version_{branch} --pretty=%s".format(prev=prev_version, branch=branch)
-    change_log = common.run_shell_command(change_log_cmd)
-
-    if change_log == "":
-        print("No changes found. Exiting...")
-        return
-
-    # Recap
-    questions = [
-        inquirer.List(
-            'choice',
-            message="We will create a release tagged {tag} from branch '{branch}'".format(tag=next_version, branch=branch),
-            choices=['Yes', 'No'],
-        ),
-    ]
-    answers = inquirer.prompt(questions)
-    if answers['choice'] == 'No':
-        print('Aborting...')
-        return
-
-    # Creates the release
-    github.create_releases(owner, app, next_version, branch, change_log, prerelease)
-    print("Done ! Check it at: https://github.com/{owner}/{repo}/releases/tag/{tag}".format(owner=owner, repo=app, tag=next_version))
-
 
 ###############################################################################
