+++ install.sh
@@ -153,6 +153,7 @@ export DMAKE_VERSION=\${DMAKE_VERSION:-${DMAKE_VERSION}}
 export DMAKE_UID=\${DMAKE_UID:-${DMAKE_UID}}
 export DMAKE_PATH=\${DMAKE_PATH:-${DMAKE_PATH}}
 export DMAKE_CONFIG_DIR=\${DMAKE_CONFIG_DIR:-${DMAKE_CONFIG_DIR}}
+export DMAKE_NO_GPU=\${DMAKE_NO_GPU:-${DMAKE_NO_GPU}}
 export DMAKE_PULL_CONFIG_DIR=\${DMAKE_PULL_CONFIG_DIR:-${DMAKE_PULL_CONFIG_DIR}}
 export DMAKE_USE_HOST_PORTS=\${DMAKE_USE_HOST_PORTS:-${DMAKE_USE_HOST_PORTS}}
 export DMAKE_SSH_KEY=\${DMAKE_SSH_KEY:-${DMAKE_SSH_KEY}}
