+++ requirements.txt
@@ -5,4 +5,5 @@ backports.ssl-match-hostname>=3.5.0.1
 ipaddress>=1.0.18
 oauthlib>=2.0.6
 requests-oauthlib==0.8.0
-git+https://github.com/magmax/python-inquirer.git
+git+https://github.com/magmax/python-inquirer.git@2.2.0
+PyGithub==1.39
